Rewrote client and manager
